fix: remove analyst constant in build
