+++ src/components/UI/molecules/UserMenuWrapperAnalyst/UserMenuWrapperAnalyst.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Meta, StoryObj } from '@storybook/react'
-import { UserMenuAnalystProps } from '@constants/stories'
+import { UserMenuAnalystProps } from '@constants/stories/analyst.constants'
 import { UserMenuWrapperAnalyst } from './UserMenuWrapperAnalyst.component'
 
 const meta: Meta<typeof UserMenuWrapperAnalyst> = {
