+++ src/components/UI/organism/ModalAnalyst/ModalAnalyst.stories.ts
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react'
-import { ModalAnalystProps } from '@constants/stories'
+import { ModalAnalystProps } from '@constants/stories/analyst.constants'
 import { ModalAnalyst } from './ModalAnalyst.component'
 
 const meta: Meta<typeof ModalAnalyst> = {
