+++ src/constants/stories/index.ts
@@ -1,6 +1,5 @@
 export * from './DatePicker.constants'
 export * from './Image.constants'
-export * from './analyst.constants'
 export * from './citiesDrawer.constant'
 export * from './common.constants'
 export * from './drawer.constants'
